+++ src/renderer/components/swap/Swap.stories.tsx
@@ -23,7 +23,6 @@ import { THORCHAIN_DECIMAL } from '../../helpers/assetHelper'
 import { RUNE_PRICE_POOL } from '../../helpers/poolHelper'
 import { INITIAL_SWAP_STATE } from '../../services/chain/const'
 import { SwapState } from '../../services/chain/types'
-import { DEFAULT_MIMIR_HALT } from '../../services/thorchain/const'
 import { Swap as Component, SwapProps } from './Swap'
 import { SwapAsset } from './Swap.types'
 
@@ -32,8 +31,7 @@ const targetAsset: SwapAsset = { asset: AssetBTC, decimal: BTC_DECIMAL, price: b
 
 /* Mock all (default) data needed by `Swap` commponent */
 const defaultProps: SwapProps = {
-  haltedChains: [],
-  mimirHalt: DEFAULT_MIMIR_HALT,
+  disableSwapAction: false,
   keystore: O.none,
   poolAssets: [AssetBTC, AssetRuneNative],
   assets: { source: sourceAsset, target: targetAsset },
@@ -103,11 +101,15 @@ const defaultProps: SwapProps = {
       })
     ),
   approveFee$: () => Rx.of(RD.success(baseAmount(10000000))),
-  sourceWalletAddress: O.some('source-wallet-address'),
+  sourceKeystoreAddress: O.some('source-wallet-address'),
   sourceLedgerAddress: O.some('source-ledger-address'),
-  targetWalletAddress: O.some('target-wallet-address'),
+  targetKeystoreAddress: O.some('target-wallet-address'),
   targetLedgerAddress: O.some('target-ledger-address'),
-  onChangePath: (path) => console.log('change path', path),
+  recipientAddress: O.some('target-address'),
+  sourceWalletType: 'ledger',
+  targetWalletType: O.some('keystore'),
+  onChangeAsset: ({ source, target, sourceWalletType, targetWalletType }) =>
+    console.log('change asset', assetToString(source), sourceWalletType, assetToString(target), targetWalletType),
   network: 'testnet',
   slipTolerance: 5,
   changeSlipTolerance: () => console.log('changeSlipTolerance'),