+++ src/renderer/components/swap/Swap.tsx
@@ -17,7 +17,6 @@ import {
   baseAmount,
   formatAssetAmountCurrency,
   delay,
-  Chain,
   assetToBase,
   assetAmount,
   chainToString,
@@ -32,6 +31,7 @@ import { useIntl } from 'react-intl'
 import * as RxOp from 'rxjs/operators'
 
 import { Network } from '../../../shared/api/types'
+import { isLedgerWallet } from '../../../shared/utils/guard'
 import { WalletType } from '../../../shared/wallet/types'
 import { ZERO_BASE_AMOUNT } from '../../const'
 import {
@@ -47,7 +47,7 @@ import {
 } from '../../helpers/assetHelper'
 import { getChainAsset, isBchChain, isBtcChain, isDogeChain, isEthChain, isLtcChain } from '../../helpers/chainHelper'
 import { unionAssets } from '../../helpers/fp/array'
-import { eqAsset, eqBaseAmount, eqOAsset, eqOApproveParams, eqAddress, eqOAddress } from '../../helpers/fp/eq'
+import { eqAsset, eqBaseAmount, eqOAsset, eqOApproveParams, eqAddress } from '../../helpers/fp/eq'
 import { sequenceSOption, sequenceTOption } from '../../helpers/fpHelpers'
 import * as PoolHelpers from '../../helpers/poolHelper'
 import { liveData, LiveData } from '../../helpers/rx/liveData'
@@ -55,10 +55,10 @@ import { emptyString, loadingString, noDataString } from '../../helpers/stringHe
 import {
   filterWalletBalancesByAssets,
   getWalletBalanceByAssetAndWalletType,
+  getWalletTypeLabel,
   hasLedgerInBalancesByAsset
 } from '../../helpers/walletHelper'
 import { useSubscriptionState } from '../../hooks/useSubscriptionState'
-import { swap } from '../../routes/pools'
 import { ChangeSlipToleranceHandler } from '../../services/app/types'
 import { INITIAL_SWAP_STATE } from '../../services/chain/const'
 import { getZeroSwapFees } from '../../services/chain/fees/swap'
@@ -81,7 +81,6 @@ import {
   LoadApproveFeeHandler
 } from '../../services/ethereum/types'
 import { PoolAddress, PoolDetails, PoolsDataMap } from '../../services/midgard/types'
-import { MimirHalt } from '../../services/thorchain/types'
 import {
   ApiError,
   KeystoreState,
@@ -127,8 +126,10 @@ export type SwapProps = {
     source: SwapAsset
     target: SwapAsset
   }
-  sourceWalletAddress: O.Option<Address>
+  sourceKeystoreAddress: O.Option<Address>
   sourceLedgerAddress: O.Option<Address>
+  sourceWalletType: WalletType
+  targetWalletType: O.Option<WalletType>
   poolAddress: O.Option<PoolAddress>
   swap$: SwapStateHandler
   poolsData: PoolsDataMap
@@ -143,17 +144,29 @@ export type SwapProps = {
   fees$: SwapFeesHandler
   reloadApproveFee: LoadApproveFeeHandler
   approveFee$: ApproveFeeHandler
-  targetWalletAddress: O.Option<Address>
+  recipientAddress: O.Option<Address>
+  targetKeystoreAddress: O.Option<Address>
   targetLedgerAddress: O.Option<Address>
-  onChangePath: (path: string) => void
+  onChangeAsset: ({
+    source,
+    sourceWalletType,
+    target,
+    targetWalletType,
+    recipientAddress
+  }: {
+    source: Asset
+    target: Asset
+    sourceWalletType: WalletType
+    targetWalletType: O.Option<WalletType>
+    recipientAddress: O.Option<Address>
+  }) => void
   network: Network
   slipTolerance: SlipTolerance
   changeSlipTolerance: ChangeSlipToleranceHandler
   approveERC20Token$: (params: ApproveParams) => TxHashLD
   isApprovedERC20Token$: (params: IsApproveParams) => LiveData<ApiError, boolean>
   importWalletHandler: FP.Lazy<void>
-  haltedChains: Chain[]
-  mimirHalt: MimirHalt
+  disableSwapAction: boolean
   clickAddressLinkHandler: (address: Address) => void
   addressValidator: AddressValidationAsync
 }
@@ -177,11 +190,14 @@ export const Swap = ({
   reloadFees,
   reloadBalances = FP.constVoid,
   fees$,
-  sourceWalletAddress: oInitialSourceWalletAddress,
+  sourceKeystoreAddress: oInitialSourceKeystoreAddress,
   sourceLedgerAddress: oSourceLedgerAddress,
-  targetWalletAddress: oInitialTargetWalletAddress,
+  targetKeystoreAddress: oTargetKeystoreAddress,
   targetLedgerAddress: oTargetLedgerAddress,
-  onChangePath,
+  recipientAddress: oRecipientAddress,
+  sourceWalletType: initialSourceWalletType,
+  targetWalletType: oInitialTargetWalletType,
+  onChangeAsset,
   network,
   slipTolerance,
   changeSlipTolerance,
@@ -190,8 +206,7 @@ export const Swap = ({
   reloadApproveFee,
   approveFee$,
   importWalletHandler,
-  haltedChains,
-  mimirHalt,
+  disableSwapAction,
   clickAddressLinkHandler,
   addressValidator
 }: SwapProps) => {
@@ -199,21 +214,21 @@ export const Swap = ({
 
   const lockedWallet: boolean = useMemo(() => isLocked(keystore) || !hasImportedKeystore(keystore), [keystore])
 
-  const [oSourceWalletAddress, setSourceWalletAddress] = useState<O.Option<Address>>(oInitialSourceWalletAddress)
-  // Update state needed - initial walletAddress is loaded async and can be different at first run
-  useEffect(() => {
-    setSourceWalletAddress(oInitialSourceWalletAddress)
-  }, [oInitialSourceWalletAddress])
+  const useSourceAssetLedger = isLedgerWallet(initialSourceWalletType)
+  const oSourceWalletAddress = useSourceAssetLedger ? oSourceLedgerAddress : oInitialSourceKeystoreAddress
 
-  const [oTargetWalletAddress, setTargetWalletAddress] = useState<O.Option<Address>>(oInitialTargetWalletAddress)
-  const [editableTargetWalletAddress, setEditableTargetWalletAddress] =
-    useState<O.Option<Address>>(oInitialTargetWalletAddress)
+  const useTargetAssetLedger = FP.pipe(
+    oInitialTargetWalletType,
+    O.map(isLedgerWallet),
+    O.getOrElse(() => false)
+  )
+
+  const [oTargetWalletType, setTargetWalletType] = useState<O.Option<WalletType>>(oInitialTargetWalletType)
 
   // Update state needed - initial target walletAddress is loaded async and can be different at first run
   useEffect(() => {
-    setTargetWalletAddress(oInitialTargetWalletAddress)
-    setEditableTargetWalletAddress(oInitialTargetWalletAddress)
-  }, [oInitialTargetWalletAddress])
+    setTargetWalletType(oInitialTargetWalletType)
+  }, [oInitialTargetWalletType])
 
   const { balances: oWalletBalances, loading: walletBalancesLoading } = walletBalances
 
@@ -226,33 +241,7 @@ export const Swap = ({
   const prevSourceAsset = useRef<O.Option<Asset>>(O.none)
   const prevTargetAsset = useRef<O.Option<Asset>>(O.none)
 
-  const [useSourceAssetLedger, setUseSourceAssetLedger] = useState(false)
-  const [useTargetAssetLedger, setUseTargetAssetLedger] = useState(false)
-
-  const oTargetAddress: O.Option<Address> = useTargetAssetLedger ? oTargetLedgerAddress : oTargetWalletAddress
-
-  const disableAllPoolActions = useCallback(
-    (chain: Chain) => PoolHelpers.disableAllActions({ chain, haltedChains, mimirHalt }),
-    [haltedChains, mimirHalt]
-  )
-  const disableTradingPoolActions = useCallback(
-    (chain: Chain) => PoolHelpers.disableTradingActions({ chain, haltedChains, mimirHalt }),
-    [haltedChains, mimirHalt]
-  )
-
   const [customAddressEditActive, setCustomAddressEditActive] = useState(false)
-
-  const disableSwapAction = useMemo(() => {
-    const sourceChain = sourceAsset.chain
-    const targetChain = targetAsset.chain
-    return (
-      disableAllPoolActions(sourceChain) ||
-      disableTradingPoolActions(sourceChain) ||
-      disableAllPoolActions(targetChain) ||
-      disableTradingPoolActions(targetChain) ||
-      customAddressEditActive
-    )
-  }, [customAddressEditActive, disableAllPoolActions, disableTradingPoolActions, sourceAsset.chain, targetAsset.chain])
 
   /**
    * All balances based on available assets to swap
@@ -275,22 +264,23 @@ export const Swap = ({
 
   const hasTargetAssetLedger = useMemo(() => O.isSome(oTargetLedgerAddress), [oTargetLedgerAddress])
 
-  const oTargetWalletType: O.Option<WalletType> = useMemo(() => {
-    // Check for Ledger
-    if (hasTargetAssetLedger && eqOAddress.equals(editableTargetWalletAddress, oTargetLedgerAddress)) {
-      return O.some('ledger')
-    }
-    // Check for keystore
-    if (
-      O.isSome(oInitialTargetWalletAddress) &&
-      eqOAddress.equals(editableTargetWalletAddress, oInitialTargetWalletAddress)
-    ) {
-      return O.some('keystore')
-    }
-    // unknown type
-    return O.none
-  }, [editableTargetWalletAddress, hasTargetAssetLedger, oInitialTargetWalletAddress, oTargetLedgerAddress])
+  const getTargetWalletTypeByAddress = useCallback(
+    (address: Address): O.Option<WalletType> => {
+      const isKeystoreAddress = FP.pipe(
+        oTargetKeystoreAddress,
+        O.map((keystoreAddress) => eqAddress.equals(keystoreAddress, address)),
+        O.getOrElse(() => false)
+      )
+      const isLedgerAddress = FP.pipe(
+        oTargetLedgerAddress,
+        O.map((ledgerAddress) => eqAddress.equals(ledgerAddress, address)),
+        O.getOrElse(() => false)
+      )
 
+      return isKeystoreAddress ? O.some('keystore') : isLedgerAddress ? O.some('ledger') : O.none
+    },
+    [oTargetLedgerAddress, oTargetKeystoreAddress]
+  )
   const sourceWalletType: WalletType = useMemo(
     () => (useSourceAssetLedger ? 'ledger' : 'keystore'),
     [useSourceAssetLedger]
@@ -689,7 +679,7 @@ export const Swap = ({
   const oSwapParams: O.Option<SwapTxParams> = useMemo(
     () =>
       FP.pipe(
-        sequenceTOption(oPoolAddress, oTargetAddress, oSourceAssetWB),
+        sequenceTOption(oPoolAddress, oRecipientAddress, oSourceAssetWB),
         O.map(([poolAddress, address, { walletType, walletAddress, walletIndex, hdMode }]) => {
           const memo = getSwapMemo({
             asset: targetAsset,
@@ -711,7 +701,7 @@ export const Swap = ({
       ),
     [
       oPoolAddress,
-      oTargetAddress,
+      oRecipientAddress,
       oSourceAssetWB,
       targetAsset,
       swapLimit1e8,
@@ -890,14 +880,16 @@ export const Swap = ({
       // delay to avoid render issues while switching
       await delay(100)
 
-      onChangePath(
-        swap.path({
-          source: assetToString(asset),
-          target: assetToString(targetAsset)
-        })
-      )
+      onChangeAsset({
+        source: asset,
+        // back to default 'keystore' type
+        sourceWalletType: 'keystore',
+        target: targetAsset,
+        targetWalletType: oTargetWalletType,
+        recipientAddress: oRecipientAddress
+      })
     },
-    [onChangePath, targetAsset]
+    [oRecipientAddress, oTargetWalletType, onChangeAsset, targetAsset]
   )
 
   const setTargetAsset = useCallback(
@@ -905,14 +897,17 @@ export const Swap = ({
       // delay to avoid render issues while switching
       await delay(100)
 
-      onChangePath(
-        swap.path({
-          source: assetToString(sourceAsset),
-          target: assetToString(asset)
-        })
-      )
+      onChangeAsset({
+        source: sourceAsset,
+        sourceWalletType,
+        target: asset,
+        // back to default 'keystore' type
+        targetWalletType: O.some('keystore'),
+        // Set recipient address to 'none' will lead to use keystore address in `WalletView`
+        recipientAddress: O.none
+      })
     },
-    [sourceAsset, onChangePath]
+    [onChangeAsset, sourceAsset, sourceWalletType]
   )
 
   const minAmountToSwapMax1e8: BaseAmount = useMemo(
@@ -1492,31 +1487,23 @@ export const Swap = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [sourceAsset, targetAsset])
 
-  /**
-   * Disables switch of assets based on following rule:
-   * Target asset (which will be the new source asset after a switch ) needs to have balances.
-   * Without balances, a source asset can't doing a swap.
-   */
-  const disableSwitchAssets = useMemo(() => {
-    const hasTargetBalance = FP.pipe(
-      oTargetAssetAmount,
-      O.map((amount) => amount.gt(baseAmount(0, targetAssetDecimal))),
-      O.getOrElse(() => false)
-    )
-    return !hasTargetBalance
-  }, [oTargetAssetAmount, targetAssetDecimal])
-
   const onSwitchAssets = useCallback(async () => {
     // delay to avoid render issues while switching
     await delay(100)
 
-    onChangePath(
-      swap.path({
-        target: assetToString(sourceAsset),
-        source: assetToString(targetAsset)
-      })
+    const walletType = FP.pipe(
+      oTargetWalletType,
+      O.getOrElse<WalletType>(() => 'keystore')
     )
-  }, [onChangePath, sourceAsset, targetAsset])
+
+    onChangeAsset({
+      source: targetAsset,
+      sourceWalletType: walletType,
+      target: sourceAsset,
+      targetWalletType: O.some(sourceWalletType),
+      recipientAddress: oSourceWalletAddress
+    })
+  }, [oSourceWalletAddress, oTargetWalletType, onChangeAsset, sourceAsset, sourceWalletType, targetAsset])
 
   const disableSubmit: boolean = useMemo(
     () =>
@@ -1530,7 +1517,8 @@ export const Swap = ({
       minAmountError ||
       isCausedSlippage ||
       swapResultAmountMax1e8.lte(zeroTargetBaseAmountMax1e8) ||
-      O.isNone(oTargetAddress),
+      O.isNone(oRecipientAddress) ||
+      customAddressEditActive,
     [
       disableSwapAction,
       lockedWallet,
@@ -1543,7 +1531,8 @@ export const Swap = ({
       isCausedSlippage,
       swapResultAmountMax1e8,
       zeroTargetBaseAmountMax1e8,
-      oTargetAddress
+      oRecipientAddress,
+      customAddressEditActive
     ]
   )
 
@@ -1553,35 +1542,53 @@ export const Swap = ({
     [checkIsApprovedError, isApproveFeeError, oApproveParams, walletBalancesLoading]
   )
 
-  const onChangeTargetAddress = useCallback(
+  const onChangeRecipientAddress = useCallback(
     (address: Address) => {
-      setTargetWalletAddress(O.some(address))
-
-      // update state of `useTargetAssetLedger`
-      const isTargetLedgerAddress = FP.pipe(
-        oTargetLedgerAddress,
-        O.map((ledgerAddress) => eqAddress.equals(ledgerAddress, address)),
-        O.getOrElse(() => false)
-      )
-      setUseTargetAssetLedger(isTargetLedgerAddress)
+      onChangeAsset({
+        source: sourceAsset,
+        target: targetAsset,
+        sourceWalletType,
+        targetWalletType: getTargetWalletTypeByAddress(address),
+        recipientAddress: O.some(address)
+      })
     },
-    [oTargetLedgerAddress]
+    [getTargetWalletTypeByAddress, onChangeAsset, sourceAsset, targetAsset, sourceWalletType]
   )
 
-  const onClickUseSourceAssetLedger = useCallback(() => {
-    const useLedger = !useSourceAssetLedger
-    setUseSourceAssetLedger(() => !useSourceAssetLedger)
-    const oAddress = useLedger ? oSourceLedgerAddress : oInitialSourceWalletAddress
-    setSourceWalletAddress(oAddress)
-  }, [oInitialSourceWalletAddress, oSourceLedgerAddress, useSourceAssetLedger])
+  const onChangeEditableRecipientAddress = useCallback(
+    (address: Address) => {
+      // Check and show wallet type while typing a custom recipient address
+      const walletType = getTargetWalletTypeByAddress(address)
+      setTargetWalletType(walletType)
+    },
+    [getTargetWalletTypeByAddress]
+  )
 
-  const onClickUseTargetAssetLedger = useCallback(() => {
-    const useLedger = !useTargetAssetLedger
-    setUseTargetAssetLedger(useLedger)
-    const oAddress = useLedger ? oTargetLedgerAddress : oInitialTargetWalletAddress
-    setTargetWalletAddress(oAddress)
-    setEditableTargetWalletAddress(oAddress)
-  }, [oInitialTargetWalletAddress, oTargetLedgerAddress, useTargetAssetLedger])
+  const onClickUseSourceAssetLedger = useCallback(
+    (useLedger: boolean) => {
+      onChangeAsset({
+        source: sourceAsset,
+        target: targetAsset,
+        sourceWalletType: useLedger ? 'ledger' : 'keystore',
+        targetWalletType: oTargetWalletType,
+        recipientAddress: oRecipientAddress
+      })
+    },
+    [oRecipientAddress, oTargetWalletType, onChangeAsset, sourceAsset, targetAsset]
+  )
+
+  const onClickUseTargetAssetLedger = useCallback(
+    (useLedger: boolean) => {
+      onChangeAsset({
+        source: sourceAsset,
+        target: targetAsset,
+        sourceWalletType,
+        targetWalletType: O.some(useLedger ? 'ledger' : 'keystore'),
+        recipientAddress: useLedger ? oTargetLedgerAddress : oTargetKeystoreAddress
+      })
+    },
+    [oTargetLedgerAddress, oTargetKeystoreAddress, onChangeAsset, sourceAsset, sourceWalletType, targetAsset]
+  )
 
   const memoTitle = useMemo(
     () =>
@@ -1684,15 +1691,8 @@ export const Swap = ({
         />
 
         <div className="mb-20px flex w-full justify-center">
-          <BaseButton
-            disabled={disableSwitchAssets}
-            onClick={!disableSwitchAssets ? () => onSwitchAssets() : undefined}
-            className="group w-full">
-            <ArrowsUpDownIcon
-              className={`ease h-[30px] w-[30px] text-turquoise
-                   ${!disableSwitchAssets ? 'group-hover:rotate-180' : ''}
-                      `}
-            />
+          <BaseButton onClick={onSwitchAssets} className="group w-full">
+            <ArrowsUpDownIcon className="ease h-[30px] w-[30px] text-turquoise group-hover:rotate-180" />
           </BaseButton>
         </div>
         <div className="flex flex-col">
@@ -1713,18 +1713,14 @@ export const Swap = ({
         </div>
         {!lockedWallet &&
           FP.pipe(
-            oTargetAddress,
+            oRecipientAddress,
             O.map((address) => (
               <div className="mt-20px flex flex-col  px-10px" key="edit-address">
                 <div className="flex items-center">
                   <h3 className="font-[12px] !mb-0 mr-10px w-auto p-0 font-main uppercase text-text2 dark:text-text2d">
                     {intl.formatMessage({ id: 'common.recipient' })}
                   </h3>
-                  {FP.pipe(
-                    oTargetWalletType,
-                    O.map((walletType) => <WalletTypeLabel key="target-w-type">{walletType}</WalletTypeLabel>),
-                    O.toNullable
-                  )}
+                  <WalletTypeLabel key="target-w-type">{getWalletTypeLabel(oTargetWalletType, intl)}</WalletTypeLabel>
                 </div>
                 <EditableAddress
                   key={address}
@@ -1732,8 +1728,8 @@ export const Swap = ({
                   network={network}
                   address={address}
                   onClickOpenAddress={(address) => clickAddressLinkHandler(address)}
-                  onChangeAddress={onChangeTargetAddress}
-                  onChangeEditableAddress={(newAddress) => setEditableTargetWalletAddress(O.some(newAddress))}
+                  onChangeAddress={onChangeRecipientAddress}
+                  onChangeEditableAddress={onChangeEditableRecipientAddress}
                   onChangeEditableMode={(editModeActive) => setCustomAddressEditActive(editModeActive)}
                   addressValidator={addressValidator}
                 />
@@ -1941,7 +1937,7 @@ export const Swap = ({
                       <div>{intl.formatMessage({ id: 'common.recipient' })}</div>
                       <div className="truncate pl-20px text-[13px] normal-case leading-normal">
                         {FP.pipe(
-                          oTargetAddress,
+                          oRecipientAddress,
                           O.map((address) => (
                             <TooltipAddress title={address} key="tooltip-target-addr">
                               {address}