+++ src/renderer/routes/pools/pools.test.ts
@@ -30,26 +30,26 @@ describe('Pools routes', () => {
 
   describe('Deposit routes', () => {
     it('template', () => {
-      expect(deposit.template).toEqual('/pools/deposit/:asset|:assetWalletType|:runeWalletType')
+      expect(deposit.template).toEqual('/pools/deposit/:asset/:assetWalletType/:runeWalletType')
     })
     it('asset - keystore | rune - keystore', () => {
       expect(deposit.path({ asset: 'BNB.BNB', assetWalletType: 'keystore', runeWalletType: 'keystore' })).toEqual(
-        '/pools/deposit/bnb.bnb|keystore|keystore'
+        '/pools/deposit/bnb.bnb/keystore/keystore'
       )
     })
     it('asset - ledger | rune - keystore', () => {
       expect(deposit.path({ asset: 'BNB.BNB', assetWalletType: 'ledger', runeWalletType: 'keystore' })).toEqual(
-        '/pools/deposit/bnb.bnb|ledger|keystore'
+        '/pools/deposit/bnb.bnb/ledger/keystore'
       )
     })
     it('asset - keystore | rune - ledger', () => {
       expect(deposit.path({ asset: 'BNB.BNB', assetWalletType: 'keystore', runeWalletType: 'ledger' })).toEqual(
-        '/pools/deposit/bnb.bnb|keystore|ledger'
+        '/pools/deposit/bnb.bnb/keystore/ledger'
       )
     })
     it('asset - ledger | rune - ledger', () => {
       expect(deposit.path({ asset: 'BNB.BNB', assetWalletType: 'ledger', runeWalletType: 'ledger' })).toEqual(
-        '/pools/deposit/bnb.bnb|ledger|ledger'
+        '/pools/deposit/bnb.bnb/ledger/ledger'
       )
     })
     it('redirects for empty assets', () => {
@@ -61,16 +61,57 @@ describe('Pools routes', () => {
 
   describe('Swap routes', () => {
     it('template', () => {
-      expect(swap.template).toEqual('/pools/swap/:source|:target')
+      expect(swap.template).toEqual('/pools/swap/:source/:sourceWalletType/:target/:targetWalletType/:recipient?')
     })
-    it('returns path by given source/target parameters', () => {
-      expect(swap.path({ source: 'BNB.BNB', target: 'THOR.RUNE' })).toEqual('/pools/swap/bnb.bnb|thor.rune')
+    it('source keystore / target keystore', () => {
+      expect(
+        swap.path({
+          source: 'BNB.BNB',
+          sourceWalletType: 'keystore',
+          target: 'THOR.RUNE',
+          targetWalletType: 'keystore'
+        })
+      ).toEqual('/pools/swap/bnb.bnb/keystore/thor.rune/keystore')
     })
-    it('redirects to base path if source is empty', () => {
-      expect(swap.path({ source: '', target: 'THOR.RUNE' })).toEqual('/pools/swap')
+    it('source ledger / target keystore', () => {
+      expect(
+        swap.path({
+          source: 'BNB.BNB',
+          sourceWalletType: 'ledger',
+          target: 'THOR.RUNE',
+          targetWalletType: 'keystore'
+        })
+      ).toEqual('/pools/swap/bnb.bnb/ledger/thor.rune/keystore')
     })
-    it('redirects to base path if target is empty', () => {
-      expect(swap.path({ source: 'BNB.BNB', target: '' })).toEqual('/pools/swap')
+    it('source keystore / target ledger', () => {
+      expect(
+        swap.path({
+          source: 'BNB.BNB',
+          sourceWalletType: 'keystore',
+          target: 'THOR.RUNE',
+          targetWalletType: 'ledger'
+        })
+      ).toEqual('/pools/swap/bnb.bnb/keystore/thor.rune/ledger')
+    })
+    it('source ledger / target ledger', () => {
+      expect(
+        swap.path({
+          source: 'BTC.BTC',
+          sourceWalletType: 'ledger',
+          target: 'ETH.ETH',
+          targetWalletType: 'ledger'
+        })
+      ).toEqual('/pools/swap/btc.btc/ledger/eth.eth/ledger')
+    })
+    it('redirects if source is empty', () => {
+      expect(
+        swap.path({ source: '', target: 'THOR.RUNE', sourceWalletType: 'keystore', targetWalletType: 'keystore' })
+      ).toEqual('/pools/swap')
+    })
+    it('redirects if target is empty', () => {
+      expect(
+        swap.path({ source: 'BNB.BNB', target: '', sourceWalletType: 'keystore', targetWalletType: 'keystore' })
+      ).toEqual('/pools/swap')
     })
   })