+++ src/renderer/routes/pools/swap.ts
@@ -1,3 +1,4 @@
+import { WalletType } from '../../../shared/wallet/types'
 import { Route } from '../types'
 import { base as poolsBase } from './base'
 
@@ -8,16 +9,28 @@ export const base: Route<void> = {
   }
 }
 
-export type SwapRouteParams = { source: string; target: string }
+export type SwapRouteTargetWalletType = WalletType | 'custom'
+export type SwapRouteParams = {
+  source: string
+  target: string
+  sourceWalletType: WalletType
+  targetWalletType: SwapRouteTargetWalletType
+  recipient?: string // Recipient address
+}
+
 export const swap: Route<SwapRouteParams> = {
   /**
    * Use '|' 'cause asset symbols have '-' separator
    * `walletType` - wallet type of the source
    */
-  template: `${base.template}/:source|:target`,
-  path: ({ source, target }) => {
+  template: `${base.template}/:source/:sourceWalletType/:target/:targetWalletType/:recipient?`,
+  path: ({ source, target, sourceWalletType, targetWalletType, recipient }) => {
     if (!!source && !!target) {
-      return `${base.template}/${source.toLowerCase()}|${target.toLowerCase()}`
+      const basePath = `${
+        base.template
+      }/${source.toLowerCase()}/${sourceWalletType}/${target.toLowerCase()}/${targetWalletType}`
+
+      return recipient ? `${basePath}/${recipient}` : basePath
     }
     // Redirect to base route if passed params are empty
     return base.path()