+++ src/renderer/views/swap/SwapView.tsx
@@ -13,12 +13,15 @@ import { useLocation, useNavigate, useParams } from 'react-router-dom'
 import * as RxOp from 'rxjs/operators'
 
 import { Network } from '../../../shared/api/types'
+import { isLedgerWallet, isWalletType } from '../../../shared/utils/guard'
+import { WalletType } from '../../../shared/wallet/types'
 import { ErrorView } from '../../components/shared/error/'
 import { Swap } from '../../components/swap'
 import { SLIP_TOLERANCE_KEY } from '../../components/swap/SelectableSlipTolerance'
 import * as Utils from '../../components/swap/Swap.utils'
 import { BackLinkButton } from '../../components/uielements/button'
 import { Button, RefreshButton } from '../../components/uielements/button'
+import { DEFAULT_WALLET_TYPE } from '../../const'
 import { useAppContext } from '../../contexts/AppContext'
 import { useChainContext } from '../../contexts/ChainContext'
 import { useEthereumContext } from '../../contexts/EthereumContext'
@@ -28,14 +31,16 @@ import { useWalletContext } from '../../contexts/WalletContext'
 import { assetInList, getAssetFromNullableString } from '../../helpers/assetHelper'
 import { eqChain, eqNetwork } from '../../helpers/fp/eq'
 import { sequenceTOption, sequenceTRD } from '../../helpers/fpHelpers'
+import * as PoolHelpers from '../../helpers/poolHelper'
 import { RUNE_PRICE_POOL } from '../../helpers/poolHelper'
-import { addressFromOptionalWalletAddress } from '../../helpers/walletHelper'
+import { addressFromOptionalWalletAddress, getWalletAddressFromNullableString } from '../../helpers/walletHelper'
 import { useMimirHalt } from '../../hooks/useMimirHalt'
 import { useNetwork } from '../../hooks/useNetwork'
 import { useOpenAddressUrl } from '../../hooks/useOpenAddressUrl'
 import { useOpenExplorerTxUrl } from '../../hooks/useOpenExplorerTxUrl'
 import { useValidateAddress } from '../../hooks/useValidateAddress'
-import { SwapRouteParams } from '../../routes/pools/swap'
+import { swap } from '../../routes/pools'
+import { SwapRouteParams, SwapRouteTargetWalletType } from '../../routes/pools/swap'
 import * as walletRoutes from '../../routes/wallet'
 import { AssetWithDecimalLD, AssetWithDecimalRD } from '../../services/chain/types'
 import { DEFAULT_SLIP_TOLERANCE } from '../../services/const'
@@ -50,9 +55,21 @@ const eqUpdateLedgerAddress = Eq.struct<UpdateLedgerAddress>({
   network: eqNetwork
 })
 
-type Props = { sourceAsset: Asset; targetAsset: Asset }
+type Props = {
+  sourceAsset: Asset
+  targetAsset: Asset
+  sourceWalletType: WalletType
+  targetWalletType: O.Option<WalletType>
+  recipientAddress: O.Option<Address>
+}
 
-const SuccessRouteView: React.FC<Props> = ({ sourceAsset, targetAsset }): JSX.Element => {
+const SuccessRouteView: React.FC<Props> = ({
+  sourceAsset,
+  targetAsset,
+  sourceWalletType,
+  targetWalletType: oTargetWalletType,
+  recipientAddress: oRecipientAddress
+}): JSX.Element => {
   const { chain: sourceChain } = sourceAsset
   const { chain: targetChain } = targetAsset
 
@@ -211,8 +228,32 @@ const SuccessRouteView: React.FC<Props> = ({ sourceAsset, targetAsset }): JSX.El
 
   const slipTolerance = useObservableState<SlipTolerance>(slipTolerance$, getStoredSlipTolerance())
 
-  const onChangePath = useCallback(
-    (path: string) => {
+  const onChangeAssetHandler = useCallback(
+    ({
+      source,
+      sourceWalletType,
+      target,
+      targetWalletType: oTargetWalletType,
+      recipientAddress: oRecipientAddress
+    }: {
+      source: Asset
+      target: Asset
+      sourceWalletType: WalletType
+      targetWalletType: O.Option<WalletType>
+      recipientAddress: O.Option<Address>
+    }) => {
+      const targetWalletType = FP.pipe(
+        oTargetWalletType,
+        O.getOrElse<SwapRouteTargetWalletType>(() => 'custom')
+      )
+      const recipient = FP.pipe(oRecipientAddress, O.toUndefined)
+      const path = swap.path({
+        source: assetToString(source),
+        sourceWalletType,
+        target: assetToString(target),
+        targetWalletType,
+        recipient
+      })
       navigate(path, { replace: true })
     },
     [navigate]
@@ -264,6 +305,19 @@ const SuccessRouteView: React.FC<Props> = ({ sourceAsset, targetAsset }): JSX.El
     updateSourceLedgerAddress$({ chain: sourceChain, network })
   }, [network, sourceChain, updateSourceLedgerAddress$])
 
+  const isTargetLedger = FP.pipe(
+    oTargetWalletType,
+    O.map(isLedgerWallet),
+    O.getOrElse(() => false)
+  )
+
+  const oRecipient: O.Option<Address> = FP.pipe(
+    oRecipientAddress,
+    O.fromPredicate(O.isSome),
+    O.flatten,
+    O.alt(() => (isTargetLedger ? oTargetLedgerAddress : oTargetKeystoreAddress))
+  )
+
   const { validateSwapAddress } = useValidateAddress(targetAssetChain)
   const openAddressUrl = useOpenAddressUrl(targetAssetChain)
 
@@ -283,7 +337,7 @@ const SuccessRouteView: React.FC<Props> = ({ sourceAsset, targetAsset }): JSX.El
           RD.fold(
             () => <></>,
             () => (
-              <div className="h-min-full flex items-center justify-center">
+              <div className="flex min-h-[600px] w-full items-center justify-center">
                 <Spin size="large" />
               </div>
             ),
@@ -312,10 +366,26 @@ const SuccessRouteView: React.FC<Props> = ({ sourceAsset, targetAsset }): JSX.El
                 A.map(({ asset }) => asset)
               )
 
+              const disableAllPoolActions = (chain: Chain) =>
+                PoolHelpers.disableAllActions({ chain, haltedChains, mimirHalt })
+
+              const disableTradingPoolActions = (chain: Chain) =>
+                PoolHelpers.disableTradingActions({ chain, haltedChains, mimirHalt })
+
+              const checkDisableSwapAction = () => {
+                const sourceChain = sourceAsset.asset.chain
+                const targetChain = targetAsset.asset.chain
+                return (
+                  disableAllPoolActions(sourceChain) ||
+                  disableTradingPoolActions(sourceChain) ||
+                  disableAllPoolActions(targetChain) ||
+                  disableTradingPoolActions(targetChain)
+                )
+              }
+
               return (
                 <Swap
-                  haltedChains={haltedChains}
-                  mimirHalt={mimirHalt}
+                  disableSwapAction={checkDisableSwapAction()}
                   keystore={keystore}
                   validatePassword$={validatePassword$}
                   goToTransaction={openExplorerTxUrl}
@@ -324,8 +394,10 @@ const SuccessRouteView: React.FC<Props> = ({ sourceAsset, targetAsset }): JSX.El
                     source: { ...sourceAsset, price: sourceAssetDetail.assetPrice },
                     target: { ...targetAsset, price: targetAssetDetail.assetPrice }
                   }}
-                  sourceWalletAddress={oSourceKeystoreAddress}
+                  sourceKeystoreAddress={oSourceKeystoreAddress}
                   sourceLedgerAddress={oSourceLedgerAddress}
+                  sourceWalletType={sourceWalletType}
+                  targetWalletType={oTargetWalletType}
                   poolAddress={selectedPoolAddress}
                   poolAssets={poolAssets}
                   poolsData={poolsData}
@@ -336,11 +408,12 @@ const SuccessRouteView: React.FC<Props> = ({ sourceAsset, targetAsset }): JSX.El
                   fees$={swapFees$}
                   reloadApproveFee={reloadApproveFee}
                   approveFee$={approveFee$}
-                  targetWalletAddress={oTargetKeystoreAddress}
+                  targetKeystoreAddress={oTargetKeystoreAddress}
                   targetLedgerAddress={oTargetLedgerAddress}
+                  recipientAddress={oRecipient}
                   swap$={swap$}
                   reloadBalances={reloadBalances}
-                  onChangePath={onChangePath}
+                  onChangeAsset={onChangeAssetHandler}
                   network={network}
                   slipTolerance={slipTolerance}
                   changeSlipTolerance={changeSlipTolerance}
@@ -360,9 +433,18 @@ const SuccessRouteView: React.FC<Props> = ({ sourceAsset, targetAsset }): JSX.El
 }
 
 export const SwapView: React.FC = (): JSX.Element => {
-  const { source, target } = useParams<SwapRouteParams>()
+  const {
+    source,
+    target,
+    sourceWalletType: routeSourceWalletType,
+    targetWalletType: routeTargetWalletType,
+    recipient
+  } = useParams<SwapRouteParams>()
   const oSourceAsset: O.Option<Asset> = useMemo(() => getAssetFromNullableString(source), [source])
   const oTargetAsset: O.Option<Asset> = useMemo(() => getAssetFromNullableString(target), [target])
+  const oRecipientAddress: O.Option<Address> = useMemo(() => getWalletAddressFromNullableString(recipient), [recipient])
+  const sourceWalletType = routeSourceWalletType || DEFAULT_WALLET_TYPE
+  const oTargetWalletType = FP.pipe(routeTargetWalletType, O.fromPredicate(isWalletType))
 
   const intl = useIntl()
 
@@ -374,12 +456,20 @@ export const SwapView: React.FC = (): JSX.Element => {
           title={intl.formatMessage(
             { id: 'routes.invalid.params' },
             {
-              params: `source: ${source}, target: ${target} `
+              params: `source: ${source}, target: ${target}`
             }
           )}
         />
       ),
-      ([sourceAsset, targetAsset]) => <SuccessRouteView sourceAsset={sourceAsset} targetAsset={targetAsset} />
+      ([sourceAsset, targetAsset]) => (
+        <SuccessRouteView
+          sourceAsset={sourceAsset}
+          targetAsset={targetAsset}
+          sourceWalletType={sourceWalletType}
+          targetWalletType={oTargetWalletType}
+          recipientAddress={oRecipientAddress}
+        />
+      )
     )
   )
 }